+++ packages/components/card/src/card.tsx
@@ -1,4 +1,3 @@
-import { CircularProgress, LinearProgress } from "@jamsr-ui/progress";
 import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { type ComponentPropsWithoutRef } from "react";
 import { cardVariants, type CardVariants } from "./style";
@@ -16,9 +15,7 @@ export const Card = <T extends React.ElementType = "div">(
     as,
     className,
     variant,
-    isPending = false,
     children,
-    isLoading = false,
     bg,
     bordered,
     ...restProps
@@ -31,17 +28,6 @@ export const Card = <T extends React.ElementType = "div">(
       className={cn(cardVariants({ variant, bg, className, bordered }))}
       {...restProps}
     >
-      {isLoading && (
-        <LinearProgress
-          className="absolute inset-0"
-          isIntermediate
-        />
-      )}
-      {isPending && (
-        <div className="absolute inset-0 z-10 grid size-full place-content-center bg-black/50">
-          <CircularProgress />
-        </div>
-      )}
       {children}
     </Component>
   );